+++ foods.html
@@ -7,7 +7,7 @@
 <body>
   <link rel="stylesheet" href="stylesheet.css" type="text/css">
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-  <script src="main.bundle.js"></script>
+  <script src="foods.bundle.js"></script>
 
   <div id="manage-foods">
     <h3 id="manage-foods-header">Manage Foods</h3>
