+++ webpack.config.js
@@ -3,6 +3,7 @@ const path = require('path');
 module.exports = {
   entry: {
     main: "./lib/index.js",
+    foods: "./lib/foods_index.js",
     test: "mocha!./test/index.js"
   },
   output: {
